perf: read Stage 2 waveform as float32, drop unused pydub dependency
diarize.py took soundfile's float64 default on a 16-bit WAV: a 1.16 GiB array
on the 2h43m reference recording, which then forced torch.from_numpy(data)
.float() to make a second full-length copy just to narrow back to float32.
Reading float32 directly halves the array to 595 MB and removes that copy.

Values are unchanged: every sample in a 16-bit WAV is int16/2**15 and exactly
representable in float32. Verified the waveform tensor pyannote receives is
bit-identical (torch.equal true, max abs diff 0.0), and a full Stage 2 run
before and after produced byte-identical segment output (37 segments on
01_Test_File_clean.wav).

pydub is removed from requirements.txt. Nothing imports it any more - Stage 1
dropped it when decoding moved to a streaming ffmpeg pass, Stage 3 when the
clean WAV moved behind _WavReader. The remaining mentions in preprocess.py and
transcribe.py are comments recording which pydub behaviour each replacement
reproduces.

KNOWN-ISSUES.md: the float64 read and the pydub dependency move to Resolved.
The remaining Stage 2 entry is rewritten, because the "VRAM OutOfMemoryError"
turns out not to be a VRAM problem. A direct probe allocating 128 MiB CUDA
blocks until failure gets 3.75 GiB with no host pressure, but fails at 1.75 GiB
with 1.47 GiB still free on the GPU when host RAM is squeezed: on Windows WDDM
a CUDA allocation needs host memory to back it, so host pressure surfaces as a
spurious CUDA OOM. Every Stage 2 OOM seen so far occurred at 0.22-1.2 GB
AvailPhys, and Stage 2 also fails with a plain 640 KB DefaultCPUAllocator
error. expandable_segments is recorded as worth one try but unlikely to help,
since there is nothing to defragment at 215 MiB allocated.

+++ stages/diarize.py
@@ -201,7 +201,13 @@ def run(
     import numpy as np
     import torch
 
-    data, sample_rate = sf.read(clean_wav_path)
+    # dtype="float32" rather than sf.read's float64 default: the clean WAV is
+    # 16-bit, so every sample is int16/2**15 and exactly representable in
+    # float32 — the values are identical either way. Reading float64 cost twice
+    # the memory (1.16 GiB vs 595 MB on a 2h43m file) and forced
+    # torch.from_numpy(data).float() below to make a second full-length copy
+    # while narrowing.
+    data, sample_rate = sf.read(clean_wav_path, dtype="float32")
     if data.ndim == 1:
         data = data[np.newaxis, :]  # add channel dim
     else:
